Updates to shell script and commenting
Consider renaming scripts to just love-maker.xxx

+++ make-bash.sh
@@ -74,26 +74,20 @@ BIN_DIR=$PROJ_ROOT/bin/$os-$arch
 if [ -e $BIN_DIR ]; then
   choice="-1"
   echo "A distribution for this platform already exists."
-  while [ $choice != "y" ] && [ $choice != "n" ]; do
+  while [ ! -z $choice ] && [ $choice != "y" ] && [ $choice != "n" ]; do
     read -p "Would you like to update it (y) or clean and update (n)? (default y/n): " choice
-    if [ -z $choice ]; then
-      choice="y"
-    fi
   done
-  if [ $choice == "n" ]; then
+  if [ ! -z $choice ] && [ $choice == "n" ]; then
     rm -rf $BIN_DIR
   fi
 fi
 
 choice="-1"
 echo "Build will start."
-while [ $choice != "y" ] && [ $choice != "n" ]; do
+while [ ! -z $choice ] && [ $choice != "y" ] && [ $choice != "n" ]; do
   read -p "Do you want to continue? (default y/n): " choice
-  if [ -z $choice ]; then
-    choice="y"
-  fi
 done
-if [ $choice == "n" ]; then
+if [ ! -z $choice ] && [ $choice == "n" ]; then
   exit
 fi
